third_party/WebKit/LayoutTests/http/tests/credentialmanager/credentialscontainer-create-with-virtual-authenticator.html

<!DOCTYPE html>
<title>Credential Manager: End-to-end tests for create() with a virtual authenticator.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="/gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="/gen/third_party/blink/public/platform/modules/credentialmanager/credential_manager.mojom.js"></script>
<script src="/gen/third_party/blink/public/platform/modules/webauthn/authenticator.mojom.js"></script>
<script src="/gen/third_party/blink/public/platform/modules/webauthn/virtual_authenticator.mojom.js"></script>
<script src="resources/test-inputs.js"></script>
<script src="resources/virtual-navigator-credentials.js"></script>
<body>
<script>

if (document.location.host != "subdomain.example.test:8443") {
  document.location = "https://subdomain.example.test:8443/credentialmanager/credentialscontainer-create-with-virtual-authenticator.html";
  promise_test(_ => new Promise(_ => {}), "Stall tests on the wrong host.");
}

promise_test(async _ => {
  let authenticators = await navigator.credentials.test.authenticators();
  assert_equals(authenticators.length, 0);
  let testAuthenticator = await navigator.credentials.test.createAuthenticator();
}, "Set up the testing environment.");

promise_test(async t => {
  let authenticators = await navigator.credentials.test.authenticators();
  assert_equals(authenticators.length, 1);
  let testAuthenticator = authenticators[0];
  assert_true(await testAuthenticator.generateAndRegisterKey(ACCEPTABLE_CREDENTIAL_ID, "subdomain.example.test"));

  var customMakeCredOptions = deepCopy(MAKE_CREDENTIAL_OPTIONS);
  customMakeCredOptions.excludeCredentials = [ACCEPTABLE_CREDENTIAL];
  return promise_rejects(t, "InvalidStateError",
    navigator.credentials.create({ publicKey : customMakeCredOptions}));
}, "navigator.credentials.create() re-registration returns InvalidStateError");

promise_test(async t => {
  var customMakeCredOptions = deepCopy(MAKE_CREDENTIAL_OPTIONS);
  customMakeCredOptions.rp.id = "";
  return promise_rejects(t, "SecurityError",
    navigator.credentials.create({ publicKey : customMakeCredOptions}));
}, "navigator.credentials.create() with empty rpId returns SecurityError");

promise_test(async t => {
  var customMakeCredOptions = deepCopy(MAKE_CREDENTIAL_OPTIONS);
  customMakeCredOptions.authenticatorSelection.requireResidentKey = true;
  return promise_rejects(t, "NotSupportedError",
    navigator.credentials.create({ publicKey : customMakeCredOptions}));
}, "navigator.credentials.create() with resident keys is disabled");

promise_test(t => {
  return navigator.credentials.test.clearAuthenticators();
}, "Clean up testing environment.");

</script>
